metadata: add duration

+++ googleapi/src/main/java/com/liskovsoft/googleapi/common/helpers/YouTubeHelper.java
@@ -107,7 +107,7 @@ public final class YouTubeHelper {
 
         String title = mediaItem.getTitle().toLowerCase();
 
-        int lengthMs = mediaItem.getDurationMs();
+        long lengthMs = mediaItem.getDurationMs();
         boolean isShortLength = lengthMs > 0 && lengthMs <= SHORTS_LEN_MS;
         return isShortLength || title.contains("#short") || title.contains("#shorts") || title.contains("#tiktok");
     }
